pyproject.toml
[project]
name = "lightrag"
version = "0.1.0"
description = "LightRAG Project"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "aioboto3>=11.3.0",
    "graspologic>=3.3.0",
    "hnswlib>=0.7.0",
    "openai>=1.3.7",
    "pika>=1.3.2",
    "prefect>=2.14.12",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.3",
    "tiktoken>=0.5.2",
    "xxhash>=3.4.1",
    "pyvis>=0.3.2",
    "neo4j>=5.0.0",
    "pymilvus>=2.3.4",
    "pymongo>=4.6.1",
    "aio_pika>=9.4.1",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "httpx>=0.26.0",
    "jwt>=1.3.1"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true